完成专题功能

+++ app/src/main/java/com/example/jiyun_training_demo/bean/TopicBean.java
@@ -61,7 +61,7 @@ public class TopicBean {
 
             private int id;
             private String title;
-            private int price_info;
+            private float price_info;
             private String scene_pic_url;
             private String subtitle;
 
@@ -81,11 +81,11 @@ public class TopicBean {
                 this.title = title;
             }
 
-            public int getPrice_info() {
+            public float getPrice_info() {
                 return price_info;
             }
 
-            public void setPrice_info(int price_info) {
+            public void setPrice_info(float price_info) {
                 this.price_info = price_info;
             }
